Fix for subrange join unreachable
Summary:
During a subrange join operation there are occasions when the subregister for
the join being considered is actually undefined.
In this case if we trace a subregister use back to a def and none exists, we
ignore it as a don't care

See http://llvm.org/PR33152

D48102 partly fixes this problem, so this commit is now rebased on top
of that, and has a new test that still fails with D48102.

Subscribers: MatzeB, qcolombet, nhaehnle

Differential Revision: https://reviews.llvm.org/D35073