Buffer memory op merging pass for s_buffer_load
This pass attempts to merge s_buffer_load_dword instructions into larger
sizes.
Constraints are that the resource descriptor, glc flag and size are all the same
for contiguous load instructions.

TODO: it should be possible to add different instructions to this as well such
as s_buffer_store and the vector variants buffer_load/buffer_store

Change-Id: Ie98107ad3b27b3c8c2da7afc3d129336f924895a