[AMDGPU] Use waterfall for readlane with non-uniform index
Readlane should have a uniform index - but in some cases it can be non-uniform.
The original implementation of readlane defaulted to using a readfirstlane of
the vgpr index on the assumption that the index would be uniform across all
lanes.

However, there are some cases where we might want a readlane like operation that
is lowered to a waterfall that performs a readlane for each distinct index value
across lanes.

Essentially we form a loop that uses readfirstlane to get the first enabled
lane's index - enable all lanes with the same index, perform a readlane and put
the result into the vgpr result register, then disable those lanes and repeat.

The pathological case will form a loop of 64 - but in most cases it will be
fewer.
In the case where the index IS uniform across all lanes, it will loop once,
which is slightly more expensive than the original code, but not a lot more.

V2: Initialize accumulator register in waterfall code

Change-Id: Ic52dfb14b053db7713b61c85e8ac766991aecdb8