[AMDGPU] Fix-up cases where writelane has 2 SGPR operands
Even though writelane doesn't have the same constraints as other valu
instructions it still can't violate the >1 SGPR operand constraint

Due to later register propagation (e.g. fixing up vgpr operands via
readfirstlane) changing writelane to only have a single SGPR is tricky.

This implementation puts a new check after SIFixSGPRCopies that prevents
multiple SGPRs being used in any writelane instructions.

The algorithm used is to check for trivial copy prop of constants into one of
the SGPR operands and perform that if possible. If this isn't possible put an
explicit copy of Src1 SGPR into M0 and use that instead (this is allowable for
writelane as the constraint is for SGPR read-port and not constant-bus access).

Change-Id: Ic3045df22db738ca6572af00138520fd28563990