[AMDGPU] Implement set of waterfall intrinsics
Implements 3 new intrinsics for implementing waterfall code for regions of code.

Waterfall is implemented as a loop that iterates over an index of values for
active lanes. For each iteration an index is picked (the first active lane) and
all lanes with the same index are left active (the rest are disabled).
The body of the code is then executed and the result accumulated into a result
vector register.
The active lane is then disabled and the next index chosen for the next
iteration.

The worst case for waterfall is one iteration per lane - but it is usually far
less than this.

The implementation uses 3 intrinsics to mark a region:
  - llvm.amdgcn.waterfall.begin
  - llvm.amdgcn.waterfall.readfirstlane
  - llvm.amdgcn.waterfall.end

The group must contain one begin and at least one readfirstlane and end
intrinsic - if the readfirstlane uses the begin index, then there will only be a
single readfirstlane and not 2.

The readfirstlane and end intrinsics are not limited to single dword values.

The waterfall loop will enclose all instructions from the begin to the final end
intrinsic.

See the test case for specific examples of use.

Change-Id: Ibac6c2f9ac1e47047eb4e0a1a616e27b048d85e6