[AMDGPU] Implement set of waterfall intrinsics
Implements 3 new intrinsics for implementing waterfall code for regions of code.

Waterfall is implemented as a loop that iterates over an index of values for
active lanes. For each iteration an index is picked (the first active lane) and
all lanes with the same index are left active (the rest are disabled).
The body of the code is then executed and the result accumulated into a result
vector register.
The active lane is then disabled and the next index chosen for the next
iteration.

The worst case for waterfall is one iteration per lane - but it is usually far
less than this.

The implementation uses 3 intrinsics to mark a region:
  - llvm.amdgcn.waterfall.begin
  - llvm.amdgcn.waterfall.readfirstlane
  - llvm.amdgcn.waterfall.end

The group must contain one begin and at least one readfirstlane and end
intrinsic - if the readfirstlane uses the begin index, then there will only be a
single readfirstlane and not 2.

The readfirstlane and end intrinsics are not limited to single dword values.

The waterfall loop will enclose all instructions from the begin to the final end
intrinsic.

See the test case for specific examples of use.

V2:
[AMDGPU] Updates to waterfall intrinsic support

Some fixes for waterfall support.
1. Intrinsics are better defined so enforce type correctness
2. Creation of the waterfall loop requires removing kill flag for operands moved
into the loop, this can otherwise cause an assert during verification
3. Fixed issue causing a tablegen failure due to float return type being used
for a scalar return value (incorrectly). The matching for intrinsic to pseudo
instruction now uses 2 template parameter types to disambiguate when using float
src types.
4. Ensure that any analyses are invalidated due to insertion of new basic
blocks. This should have been the default anyway.
5. Updated test in light of these changes. These are good exemplars for
implementors to check against.

V3:
[AMDGCN] Fixed waterfall intrinsic groups for multiple groups per BB

For cases with more than one waterfall group per basic block, the implementation
would go wrong after creating the first waterfall loop (incorrectly tracking
current BB).

Also added a test case.

V4:
[AMDGCN] Waterfall enhancements

Extra waterfall test for multiple readfirstlane intrinsics. Make sure that the
waterfall intrinsic clauses support multiple readfirstlane intrinsics.

Extended support for begin to allow multi-dword indices for the waterfall
loop. This allows the use of multiple indexes (and multiple non-uniform values)
in the same waterfall loop by combining the individual indices into a single
multi-dword index.
This has the same worst-case of 64 iterations (wave size) as before.

Added a new test case to demonstrate using multi-dword indices.

Added support for i16/f16

Changed the implementation to prevent CSE from merging clauses (by tagging the
intrinsics as having side effects and enhancing uniform detection to work in
this case as well).
Left in some support added to partially support EarlyCSE merging as it is still
valid, but probably won't be triggered much now merging is disabled.

Added support for a last.use intrinsic. See the test for an example of
this. This works in the same way as the end intrinsic, but you tag a last_use
instead (so it includes the next use in the waterfall loop). More than one use
is acceptable as long as it is in the same BB. All instructions up to the last
last.use will be included in the waterfall loop.

An important rule for waterfall clauses (that isn't detected if violated) is
that you can't have something inside the loop that is dependent on an end
intrinsic in the loop - this restriction is logically consistent - you have to use
two loops instead.

Change-Id: Ibac6c2f9ac1e47047eb4e0a1a616e27b048d85e6