[AMDGPU] Add support for TFE/LWE in image intrinsics
TFE and LWE support requires extra result registers that are written in the
event of a failure in order to detect that failure case.
The specific use-case that initiated these changes is sparse texture support.

This means that if image intrinsics are used with either option turned on, the
programmer must ensure that the return type can contain all of the expected
results. This can result in redundant registers since the vector size must be a
power-of-2.

This change takes roughly 3 parts:
1. Modify the instruction defs in tablegen to add new instruction variants that
can accomodate the extra return values as well as defining appropriate patterns
for dag selection with tfe and/or lwe enabled.
2. Extra verification code to catch cases where intrinsics have been used but
insufficient return registers are used.
3. An extra pass to zero initialize the error value return - this is because if
the error does not occur, the register is not written and thus must be zeroed
before use.

Tests have been added/modified to test the new behaviour.

Change-Id: Ifadba25b9df43a5282e3cf7778e9c087d67dd981