[AMDGPU] Handle void* case for isLegalAddressingMode
Loop strength reduction was creating a new pointer of type void* (for two
similar but different pointers to vector types) which it does
conservatively. This then provoked an assert when isLegalAddressingMode was
invoked due to void* not being handled.

Add a check which makes the conservative assumption for void* types.

Change-Id: I6715c6bc2d81bae07a427274ef94310ac6e42192