[AMDGPU] Fixed non-uniform addr64 MUBUF in shader
Summary:
SIInstrInfo::legalizeOperands was assuming that it never sees an addr64
MUBUF in a shader. But it is generated from a global load, which LLPC
uses for a load from a read-only global variable.

Fixed to detect the addr64 case, and then legalize in the same way as in
compute.

Subscribers: arsenm, kzhuravl, wdng, nhaehnle, yaxunl, dstuttard, t-tye, llvm-commits

Differential Revision: https://reviews.llvm.org/D47194

Change-Id: I8aee2a0dc37d2061de445c30c7510e40b3e52f3c