Set FP RoundToZero for interp f16 instructions
The f16 interpolation instructions require the rounding mode to be
set to Round to Zero. This change sets that mode before each f16
interp instruction, and resets it to Round to Nearest after each
f16 interp instruction.

Note that the resulting code is non-optimal, but will be optimized
by a planned pass in the future.

Change-Id: Ic2c20dac9152a454bae5f82f907c0f8d5a3a4db5