StackSlotColoring: Decide colors per stack ID
(From Matt)

I thought I fixed this in r308673, but that fix was
very broken. The assumption that any frame index can be used
in place of another was more widespread than I realized.
Even when stack slot sharing was disabled, this was still
replacing frame index uses with a different ID with a different
stack slot.

Really fix this by doing the coloring per-stack ID, so all of
the coloring logically done in a separate namespace. This is a lot
simpler than trying to figure out how to change the color if
the stack ID is different.

Differential Revision: https://reviews.llvm.org/D45968

Change-Id: I6f7402645ac5a47930feea3b8b24fee32588b0db