[TableGen] FixedLenDecoderEmitter: allow for dummy operand in MCInst
Summary:
Normally a "dummy" operand, one named in InOperandList or OutOperandList
but with no bits in the instruction encoding, is not included in an
MCInst, and the autogenerated decoding code for the disassembler
generated by FixedLenDecodeEmitter reflects that: for any such dummy
operand, no MCOperand is pushed onto the MCInst.

I have a case where it would help commoning up of essentially the same
instruction between different variants of the cpu to have a dummy
operand represented in the MCInst.

Thus I have added a new field in an instruction "hasDummyOperands".
Normally 0, when set to 1 it causes the disassembler to push an
immediate zero MCOperand for any dummy operand.

There is no test for this, but a future AMDGPU commit will use it and
thus exercise it in its tests.

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D49027

Change-Id: I82b7e1f75b466f24889434c1733c6dd78ec91dbd