[AMDGPU] Remove unnecessary Mode register writes
A previous change inserted mode register writes required to set the
rounding mode for 16 bit interpolation instructions. This was done
as part of instruction selection, which resulted in unnecessary
mode register writes.
This commit changes the approach such that teh mode changes are
added by a separate pass, which results in more optimal mode
register writes.

Change-Id: I25a7fd15cf535e8ae5ecf68f934d103af092fa07