[AMDGCN] Waterfall enhancements
Extra waterfall test for multiple readfirstlane intrinsics. Make sure that the
waterfall intrinsic clauses support multiple readfirstlane intrinsics.

Extended support for begin to allow multi-dword indices for the waterfall
loop. This allows the use of multiple indexes (and multiple non-uniform values)
in the same waterfall loop by combining the individual indices into a single
multi-dword index.
This has the same worst-case of 64 iterations (wave size) as before.

Added a new test case to demonstrate using multi-dword indices.

Added support for i16/f16

Changed the implementation to prevent CSE from merging clauses (by tagging the
intrinsics as having side effects and enhancing uniform detection to work in
this case as well).
Left in some support added to partially support EarlyCSE merging as it is still
valid, but probably won't be triggered much now merging is disabled.

Added support for a last.use intrinsic. See the test for an example of
this. This works in the same way as the end intrinsic, but you tag a last_use
instead (so it includes the next use in the waterfall loop). More than one use
is acceptable as long as it is in the same BB. All instructions up to the last
last.use will be included in the waterfall loop.

An important rule for waterfall clauses (that isn't detected if violated) is
that you can't have something inside the loop that is dependent on an end
intrinsic in the loop - this restriction is logically consistent - you have to use
two loops instead.

Change-Id: I1ef98fb6745d3259109cf971dc2b6f9d0518a067