[AMDGCN] Fix for InstCombine of image intrinsics using TFE/LWE
InstCombine simplification of memory intrinsics for AMDGCN can't handle anything
using TFE or LWE (which requires more registers in the destination to handle an
extra status flag).

This change fixes an earlier one which had a bug in the implementation causing
the wrong argument to be checked for TFE or LWE being enabled.

Change-Id: I1ce2ea9602e9b1a2b726e04310d75058651c4fd9