[AMDGPU] Avoid image dmask=0
For an image intrinsic with tfe where all result channels except tfe
were unused, I was getting an image instruction with dmask=0 and only a
single vgpr result for tfe. That is incorrect because the hardware
assumes there is at least one vgpr result, plus the one for tfe.

Fixed by forcing dmask to 1, which gives the desired two vgpr result
with tfe in the second one.

This fix should be folded into D48826 Add support for TFE/LWE.

Change-Id: Ifb4da5e70423fb6ab32b0218fba596dd4ce3101d