[RegisterCoalescer] Extend subranges to uses after join
Summary:
The test shows a case where a full use is not in a subrange because the
subreg is undefined at the use point, but, after a coalesce, the
subrange incorrectly still does not include the use even though the
subreg is now defined at that point. The incorrect live range causes a
"subrange join unreachable" assert on a later coalesce.

This commit ensures that a subrange is extended to all uses that can be
reached from a definition.

V2: Completely different fix, instead of working round the problem in
the later coalesce that actually asserted.
V5: Fixed to not extend liveness to an undef use. Spun second test out
into its own D51257 as it is a completely different problem with the
same "subrange join unreachable" symptom.
V6: Ignore debug uses.
V7: Set up undefs correctly, to avoid generating invalid subranges.

Differential Revision: https://reviews.llvm.org/D49097

Change-Id: I172cfe16d360690e921ebe606d3e90dd1cdd1b71