[RegisterCoalescer] Fixup "Fixed inconsistent followCopyChain with subreg"
Summary:
This fixes up my earlier fix D49535 rL338070.

In followCopyChain, the comparison I added to check that two subranges
arrived back at the same def point was bogus. Fixed here.

The test fails with a "Couldn't join subrange" assert without this fix,
because followCopyChain gives the wrong answer and makes a coalesce of A
and B appear impossible where there are copies A->C->B and A is a subreg
def.

Subscribers: MatzeB, qcolombet, jvesely, nhaehnle, llvm-commits

Differential Revision: https://reviews.llvm.org/D51848

Change-Id: I4257fb93800d3834d2b1894359d778f3fdb43f40