Fix test failures in SIModeRegister pass
A number of vk_example tests failed when the previous temporary
work-around was backed out, even though the mode register pass
inserted appropriate mode changes.

It was determined that when the setreg that changes the mode
register is the first instruction of the shader there is a hazard
if the mode register was modified as part of the wave setup. This
was the case in the test failures.

The fix made is to move the setreg to be later in the shader,
thereby avoiding the hazard. For PAL shaders this is possible as
there are always other instructions present that can precede the
seteg. However this is a general corner-case that may arise in
other situations where hazards may be present upon entry to a
shader.

Also corrected which instructions are marked with the FPDPRounding
flag, and made corresponding changes to the LIT test.

Change-Id: I78c0c59b614f0ac82a6d490d6571460e1438c98d