[AMDGPU] Remove explicit mode change for 16 bit interp
The 16 bit interpolation instructions require the floating
point rounding mode to be set to round-to-zero. This was
done with explicit mode changes immediately before and
after each affected instruction as a temporary workaround.

Now that the mode register pass is in place to make the mode
register changes in a more optimal way we no longer need the
workaround, and this change removes it.

Change-Id: I9ecc3978a69009dc3ec7022c1e24396a6d5eb88d