AMDGPU/InsertWaitcnts: Update VGPR/SGPR bounds when brackets are merged
Fix an issue where VGPR/SGPR bounds are not properly extended when brackets are merged.
This manifests as missing waitcnt insertions when multiple brackets are forwarded to a
successor block and the first forward has lower VGPR/SGPR bounds.

Irreducible loop test has been extended based on a CTS failure detected for GFX9.

Change-Id: I252ca052eda13419305a3e8c35c5f82ba2b2b47a