Do not force skip over export instruction in PS
This is a temporary fix to avoid a breakage where the export(s) for a
pixel shader are in control flow (due to a kill). A pixel shader must
execute an export done, even with exec=0.

A better fix would be to have a pixel-shader-only pass that inserts a
null export done in the final basic block if there is not one already
there.

Change-Id: I22f201f95d52ff699aa8cac26bb019717b20432e