[AMDGPU] Handle unexpected uniform inputs in waterfall intrinsics
Waterfall intrinsic input operands can sometimes be uniform. The waterfall
handling pass should deal with this gracefully, even though this could be
sub-optimal.

This is an initial fix for a problem observed in a game.
Subsequent better fix being worked on which should produce optimal code (by
removing redundant waterfall loops entirely if all the elements turn out to be
uniform anyway).

Change-Id: I725041da329f2173867832cf54bbf9418d3bfef6