[AMDGPU] Improved handling of uniform waterfalls
This builds on a previous fix for uniform values being specified in waterfall
loop intrinsic groups.
Previous fix would just remove any waterfall.readfirstlane intrinsics if the
input was uniform.

This fix improves on this by spotting when ALL waterfall.readfirstlanes are
uniform and removing the waterfall loop entirely.

The tests have been extended to test this mode.

The update also handles the case where not all waterfall.readfirstlanes are
removed. This is also covered in the associated tests.

Change-Id: I29c1cef3e6ff38ac8ee9145b6380ff0bdd18499c