[ConstantFolding] Fix assertion failure on non-power-of-two vector load.
The test case does an (out of bounds) load from a global constant with
type <3 x float>. InstSimplify tried to turn this into an integer load
of the whole alloc size of the vector, which is 128 bits due to
alignment padding, and then bitcast this to <3 x vector> which failed
an assertion due to the type size mismatch.

The fix is to do an integer load of the normal size of the vector, with
no alignment padding.

Change-Id: I911d744b6f048fb18df5814dfef21a9d8259aeb6