[AMDGPU] Fix for branch offset hardware workaround
(From Ryan)

This fixes a hardware bug that makes a branch offset of 0x3f unsafe.
This replaces the 32 bit branch with offset 0x3f to a 64 bit
instruction that includes the same 32 bit branch and the encoding
for a s_nop 0 to follow. The relaxer than modifies the offsets
accordingly.

Differential Revision: https://reviews.llvm.org/D63494

Change-Id: I10b7aed99d651f8159401b01bb421f105fa6288e