[AMDGPU] Allow any value in unused src0 field in v_nop
SC puts 0x80 (literal 0) in the unused src0 operand of v_nop, which made
the LLVM disassembler not recognize it. This commit makes it recognize a
v_nop with any value in the field, like s_getpc already does.

I am pushing this here for review just so I can mention SC. If it is
uncontroversial here, I will sanitize the commit message and push to
phab for upstreaming, rather than actually submitting here.

Change-Id: Iaf0363eae26ff92fc4ebc716216476adbff37a6f