Update phis in AMDGPUUnifyDivergentExitNodes
(From Steven Perron)

Summary:
The pass AMDGPUUnifyDivergentExitNodes does not update the phi nodes in the successors of blocks that is splits.  This is fixed by calling `BasicBlock::splitBasicBlock` to split the block instead of doing it manually.  This does extra work because a new conditional branch is created in `BB` which is immediately replaced, but I think the simplicity is worth it.  It also helps make the code more future proof in case other things need to be updated.

Reviewers: arsenm

Reviewed By: arsenm

Subscribers: arsenm, kzhuravl, jvesely, wdng, nhaehnle, yaxunl, dstuttard, tpr, t-tye, dnovillo

Differential Revision: https://reviews.llvm.org/D63659

Change-Id: I1719cfa45e7bc1b5307d4475da2bad382e0a0a2f