[AMDGPU] Enable v4f16 and above for v_pk_fma instructions
If isel is presented with <2 x half> vectors then it will correctly select
v_pk_fma style instructions.
If isel is presented with e.g. <4 x half> vectors it will scalarize, unlike for
other instruction types (such as fadd, fmul etc.)

Added extra support to enable this. Updated one of the tests to include a test
for this (as well as extending the test to GFX9)

Change-Id: I6992c8ef6461a1065413237ac0f31451b132bb67