Revert "[AMDGPU] Enable merging m0 initializations."
This reverts D64766.

That change caused me a problem where an LDS instruction was missing its
m0 initialization on gfx7 on a Vulkan CTS test on the GPUOpen driver.
Compiling the same code with llc on the upstream LLVM that our Vulkan
branch is based on did not show the problem. This needs further
investigation.

Change-Id: I94b380f60b074a117075ae12d15ae4babbc8c7af