AArch64: avoid creating cycle in DAG for post-increment NEON ops.
Inserting a value into Visited has the effect of terminating a search for
predecessors if that node is seen. This is legitimate for the base address, and
acts as a slight performance optimization, but the vector-building node can be
paert of a legitimate cycle so we shouldn't stop searching there.

PR43056.

git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@370036 91177308-0d34-0410-b5e6-96231b3b80d8